Move ui package to where it's used and makes more sense.
