+++ pkg/odo/cli/service/ui/ui.go
@@ -5,9 +5,7 @@ import (
 	"fmt"
 	"github.com/golang/glog"
 	scv1beta1 "github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
-	"github.com/mgutz/ansi"
 	"github.com/redhat-developer/odo/pkg/service"
-	terminal2 "golang.org/x/crypto/ssh/terminal"
 	"gopkg.in/AlecAivazis/survey.v1"
 	"gopkg.in/AlecAivazis/survey.v1/core"
 	"gopkg.in/AlecAivazis/survey.v1/terminal"
@@ -17,10 +15,8 @@ import (
 	"strings"
 )
 
-const (
-	defaultIntegerValidatorKey    = "odo_default_integer"
-	defaultColumnNumberBeforeWrap = 80
-)
+const defaultRequiredValidatorKey = "odo_default_required"
+const defaultIntegerValidatorKey = "odo_default_integer"
 
 // Validator is a function that validates that the provided interface is conform to expectations or return an error
 type Validator func(interface{}) error
@@ -42,7 +38,6 @@ func getServiceClassesCategories(categories map[string][]scv1beta1.ClusterServic
 	return keys
 }
 
-// GetServicePlanNames returns the service plan names included in the specified map
 func GetServicePlanNames(stringMap map[string]scv1beta1.ClusterServicePlan) (keys []string) {
 	keys = make([]string, len(stringMap))
 
@@ -57,7 +52,6 @@ func GetServicePlanNames(stringMap map[string]scv1beta1.ClusterServicePlan) (key
 	return keys
 }
 
-// getServiceClassMap converts the specified array of service classes to a name-service class map
 func getServiceClassMap(classes []scv1beta1.ClusterServiceClass) (classMap map[string]scv1beta1.ClusterServiceClass) {
 	classMap = make(map[string]scv1beta1.ClusterServiceClass, len(classes))
 	for _, v := range classes {
@@ -67,7 +61,6 @@ func getServiceClassMap(classes []scv1beta1.ClusterServiceClass) (classMap map[s
 	return classMap
 }
 
-// getServiceClassNames retrieves the keys (service class names) of the specified name-service class mappings
 func getServiceClassNames(stringMap map[string]scv1beta1.ClusterServiceClass) (keys []string) {
 	keys = make([]string, len(stringMap))
 
@@ -82,11 +75,10 @@ func getServiceClassNames(stringMap map[string]scv1beta1.ClusterServiceClass) (k
 	return keys
 }
 
-// handleError handles UI-related errors, in particular useful to gracefully handle ctrl-c interrupts gracefully
 func handleError(err error) {
 	if err != nil {
 		if err == terminal.InterruptErr {
-			os.Exit(1)
+			os.Exit(-1)
 		} else {
 			glog.V(4).Infof("Encountered an error processing prompt: %v", err)
 		}
@@ -126,137 +118,50 @@ func SelectClassInteractively(classesByCategory map[string][]scv1beta1.ClusterSe
 		Message: "Which kind of service do you wish to create",
 		Options: getServiceClassesCategories(classesByCategory),
 	}
-	err := survey.AskOne(prompt, &category, survey.Required)
+	err := survey.AskOne(prompt, &category, nil)
 	handleError(err)
 
 	classes := getServiceClassMap(classesByCategory[category])
-
-	// make a new displayClassInfo function available to survey templates to be able to add class information to the display
-	displayClassInfo := "displayClassInfo"
-	core.TemplateFuncs[displayClassInfo] = func(index int, pageEntries []string) string {
-		if index >= 0 && len(pageEntries) > index {
-			selected := pageEntries[index]
-			class := classes[selected]
-			return ansi.ColorCode("default+bu") + "Service class details" + ansi.ColorCode("reset") + ":\n" +
-				classInfoItem("Name", class.GetExternalName()) +
-				classInfoItem("Description", class.GetDescription()) +
-				classInfoItem("Long", getLongDescription(class))
-		}
-		return "No matching entry"
+	core.TemplateFuncs["foo"] = func(index int, pageEntries []string) string {
+		selected := pageEntries[index]
+		class := classes[selected]
+		return fmt.Sprintf("Name: %s\nDescription: %s\nLong: %s", class.GetExternalName(), class.GetDescription(), getLongDescription(class))
 	}
-	defer delete(core.TemplateFuncs, displayClassInfo)
 
-	// record original template and defer restoring it once done
-	original := survey.SelectQuestionTemplate
-	defer restoreOriginalTemplate(original)
-
-	// add more information about the currently selected class
-	survey.SelectQuestionTemplate = original + `
-	{{- if not .ShowAnswer}}
-	{{$classInfo:=(displayClassInfo .SelectedIndex .PageEntries)}}
-	  {{- if $classInfo}}
-{{$classInfo}}
-	  {{- end}}
-	{{- end}}`
+	survey.SelectQuestionTemplate = survey.SelectQuestionTemplate + `
+===
+{{- if not .ShowAnswer}}
+	{{(foo .SelectedIndex .PageEntries)}}
+{{- end}}`
 
 	prompt = &survey.Select{
 		Message: "Which " + category + " service class should we use",
 		Options: getServiceClassNames(classes),
 	}
 
-	err = survey.AskOne(prompt, &serviceType, survey.Required)
+	err = survey.AskOne(prompt, &serviceType, nil)
 	handleError(err)
 
 	return classes[serviceType], serviceType
 }
 
-// classInfoItem computes how a given service class information item should be displayed
-func classInfoItem(name, value string) string {
-	// wrap value if needed accounting for size of value "header" (its name)
-	value = wrapIfNeeded(value, len(name)+3)
-
-	if len(value) > 0 {
-		// display the name using the default color, in bold and then reset style right after
-		return ansi.ColorCode("default+b") + name + ansi.ColorCode("reset") + ": " + value + "\n"
-	}
-	return ""
-}
-
-// wrapIfNeeded wraps the given string taking the given prefix size into account based on the width of the terminal (or
-// defaultColumnNumberBeforeWrap if terminal size cannot be determined).
-func wrapIfNeeded(value string, prefixSize int) string {
-	// get the width of the terminal
-	width, _, err := terminal2.GetSize(0)
-	if width == 0 || err != nil {
-		// if for some reason we couldn't get the size use default value
-		width = defaultColumnNumberBeforeWrap
-	}
-
-	// if the value length is greater than the width, wrap it
-	// note that we need to account for the size of the name of the value being displayed before the value (i.e. its name)
-	valueSize := len(value)
-	if valueSize+prefixSize >= width {
-		// look at each line of the value
-		split := strings.Split(value, "\n")
-		for index, line := range split {
-			// for each line, trim it and split it in space-separated clusters ("words")
-			line = strings.TrimSpace(line)
-			words := strings.Split(line, " ")
-			newLine := ""
-			lineSize := 0
-
-			for _, word := range words {
-				if lineSize+len(word)+1+prefixSize < width {
-					// concatenate word to the new computed line only if adding it to the line won't make it larger than acceptable
-					newLine = newLine + " " + word
-					lineSize = lineSize + 1 + len(word) // accumulate the line size
-				} else {
-					// otherwise, break the line and add the word on a new "line"
-					newLine = newLine + "\n" + word
-					lineSize = len(word) // reset the line size
-				}
-			}
-			// replace the initial line with the new computed version
-			split[index] = strings.TrimSpace(newLine)
-		}
-		// compute the new value by joining all the modified lines
-		value = strings.Join(split, "\n")
-	}
-	return value
-}
-
-// restoreOriginalTemplate restores the original survey template once we're done with the display
-func restoreOriginalTemplate(original string) {
-	survey.SelectQuestionTemplate = original
-}
-
 // Convert the provided ClusterServiceClass to its UI representation
 func getLongDescription(class scv1beta1.ClusterServiceClass) (longDescription string) {
-	extension := class.Spec.ExternalMetadata
-	if extension != nil {
-		var meta map[string]interface{}
-		err := json.Unmarshal(extension.Raw, &meta)
-		if err != nil {
-			glog.V(4).Infof("Unable unmarshal Extension metadata for ClusterServiceClass '%v'", class.Spec.ExternalName)
-		}
-		if val, ok := meta["longDescription"]; ok {
-			longDescription = val.(string)
-		}
+	var meta map[string]interface{}
+	err := json.Unmarshal(class.Spec.ExternalMetadata.Raw, &meta)
+	if err != nil {
+		glog.V(4).Infof("Unable unmarshal Extension metadata for ClusterServiceClass '%v'", class.Spec.ExternalName)
 	}
-
+	if val, ok := meta["longDescription"]; ok {
+		longDescription = val.(string)
+	}
 	return
 }
 
 // EnterServicePropertiesInteractively lets the user enter the properties specified by the provided plan if not already
 // specified by the passed values
 func EnterServicePropertiesInteractively(svcPlan scv1beta1.ClusterServicePlan) (values map[string]string) {
-	return enterServicePropertiesInteractively(svcPlan)
-}
-
-// enterServicePropertiesInteractively lets user enter the properties interactively using the specified Stdio instance (useful
-// for testing purposes)
-func enterServicePropertiesInteractively(svcPlan scv1beta1.ClusterServicePlan, stdio ...terminal.Stdio) (values map[string]string) {
-	planDetails, _ := service.NewServicePlan(svcPlan)
+	planDetails, _ := service.NewServicePlans(svcPlan)
 
 	properties := make(map[string]service.ServicePlanParameter, len(planDetails.Parameters))
 	for _, v := range planDetails.Parameters {
@@ -265,14 +170,12 @@ func enterServicePropertiesInteractively(svcPlan scv1beta1.ClusterServicePlan, s
 
 	values = make(map[string]string, len(properties))
 
-	sort.Sort(planDetails.Parameters)
-
 	// first deal with required properties
-	for _, prop := range planDetails.Parameters {
+	for name, prop := range properties {
 		if prop.Required {
-			addValueFor(prop, values, stdio...)
+			addValueFor(prop, values)
 			// remove property from list of properties to consider
-			delete(properties, prop.Name)
+			delete(properties, name)
 		}
 	}
 
@@ -282,17 +185,12 @@ func enterServicePropertiesInteractively(svcPlan scv1beta1.ClusterServicePlan, s
 		confirm := &survey.Confirm{
 			Message: "Provide values for non-required properties",
 		}
-
-		if len(stdio) == 1 {
-			confirm.WithStdio(stdio[0])
-		}
-
-		err := survey.AskOne(confirm, &fillOptionalProps, survey.Required)
+		err := survey.AskOne(confirm, &fillOptionalProps, nil)
 		handleError(err)
 		if fillOptionalProps {
 
 			for _, prop := range properties {
-				addValueFor(prop, values, stdio...)
+				addValueFor(prop, values)
 			}
 		}
 	}
@@ -300,49 +198,50 @@ func enterServicePropertiesInteractively(svcPlan scv1beta1.ClusterServicePlan, s
 	return values
 }
 
-// always validates
-var nilValidator = func(ans interface{}) error { return nil }
-
-// getValidatorFor retrieves a Validator able to validate the specified property
-func getValidatorFor(prop service.ServicePlanParameter) (validator survey.Validator) {
-	// make sure we don't run into issues when composing validators
-	validator = nilValidator
-
-	if prop.Required {
-		validator = survey.Required
-	}
-
-	// TODO: improve, it should be possible to register validators that get chosen automatically based on passed prop
-	//  (using acceptance criteria)
-	switch prop.Type {
-	case "integer":
-		validator = survey.ComposeValidators(validator, survey.Validator(validators[defaultIntegerValidatorKey]))
-	}
-
-	return
+type chainedValidator struct {
+	validators []Validator
 }
 
-// addValueFor adds the property value based on user input to the specified values map
-func addValueFor(prop service.ServicePlanParameter, values map[string]string, stdio ...terminal.Stdio) {
+func (cv chainedValidator) validate(input interface{}) error {
+	for _, v := range cv.validators {
+		err := v(input)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func getValidatorFor(prop service.ServicePlanParameter) Validator {
+	cv := chainedValidator{}
+	if prop.Required {
+		cv.validators = append(cv.validators, validators[defaultRequiredValidatorKey])
+	}
+
+	switch prop.Type {
+	case "integer":
+		cv.validators = append(cv.validators, validators[defaultIntegerValidatorKey])
+	}
+
+	return cv.validate
+}
+
+func addValueFor(prop service.ServicePlanParameter, values map[string]string) {
 	var result string
 	prompt := &survey.Input{
 		Message: fmt.Sprintf("Enter a value for %s property %s:", prop.Type, propDesc(prop)),
-	}
-
-	if len(stdio) == 1 {
-		prompt.WithStdio(stdio[0])
 	}
 
 	if prop.HasDefaultValue {
 		prompt.Default = prop.Default
 	}
 
-	err := survey.AskOne(prompt, &result, getValidatorFor(prop))
+	err := survey.AskOne(prompt, &result, survey.Validator(getValidatorFor(prop)))
 	handleError(err)
 	values[prop.Name] = result
 }
 
-// propDesc computes a human-readable description of the specified property
 func propDesc(prop service.ServicePlanParameter) string {
 	msg := ""
 	if len(prop.Title) > 0 {
@@ -358,8 +257,9 @@ func propDesc(prop service.ServicePlanParameter) string {
 	return prop.Name + msg
 }
 
-// init initializes default validators
 func init() {
+	validators[defaultRequiredValidatorKey] = survey.Required
+
 	validators[defaultIntegerValidatorKey] = func(ans interface{}) error {
 		s := ans.(string)
 		_, err := strconv.Atoi(s)
